refactor: CardWrapper => LinkField 변경 및 UI 로직 분리
