+++ components/LinkField/LinkField.tsx
@@ -2,29 +2,29 @@
 
 import { useState } from "react";
 
-import Card from "@/components/Card/Card";
+import LinkCard from "@/components/LinkCard/LinkCard";
 import { ILink } from "@/types/linkbrary";
 
-import styles from "./CardWrapper.module.scss";
+import styles from "./LinkField.module.scss";
 
-interface ICardWrapperProps {
+interface ILinkFieldProps {
   links: ILink[];
 }
 
-const CardWrapper = ({ links }: ICardWrapperProps) => {
+const LinkField = ({ links }: ILinkFieldProps) => {
   const [openKebabIndex, setOpenKebabIndex] = useState<number>(-1);
 
   if (links.length === 0) {
     return (
-      <div className={styles.nullCardWrapper}>
+      <div className={styles.nullLinkField}>
         <p>저장된 링크가 없습니다.</p>
       </div>
     );
   } else
     return (
-      <div className={styles.cardWrapper}>
+      <div className={styles.linkField}>
         {links.map((link, index) => (
-          <Card
+          <LinkCard
             key={link.id}
             link={link}
             isClickedKebab={index === openKebabIndex}
@@ -36,4 +36,4 @@ const CardWrapper = ({ links }: ICardWrapperProps) => {
     );
 };
 
-export default CardWrapper;
+export default LinkField;
